Use etree Elements as standard
In other words, all transport/voevent messages are deserialized to an etree
Element on receipt, and all processing is done on them in that form. If an
event is forwarded, or a new transport message is generated, it is constructed
as an etree Element and then serialized to XML on transmission.