All protocols should watch for crazy lengths.
EventHandler only deals with protocols which expect to receive VOEvents. But
other protocols may also receive strings as part of the vTCP protocol, and
therefore should check for bad lengths. Everything is a subclass of
ElementSender so we implement this there.

Note that we need to take care of the MRO for old-style classes. This means we
need to inherit from ElementSender before EventHandler in class definitions.