Use SHA-1 hashes to disambiguate events
Rather than the IVORNs used previously.

Consensus in the IVOA VOEvent WG seems to be that different serializations of
the same event should share the same IVORN. If we disambiguate based on IVORN,
then eg VOEvent v1.1 and v2.0 packets describing the same event will be
regarded as the same. Since Comet will only accept each event once, it will
*either* accept the v1.1 serialization or the v2.0 serialization, but *not*
both.

It is likely that downstream clients can process only v1.1 or v2.0 events. In
the above scenario, then, Comet could cause a v1.1-only client to miss events
by forwarding only the v2.0 serialization, or vice versa.

In this revision, we change to disambiguation using SHA-1 hashes. In this
case, events are regarded as duplicates only if they are bit-for-bit
identical.

Note that this is also dangerous behaviour: if a broker changes the
composition of an event before forwarding it (something as simple as, say,
adding a trailing new-line), it will be regarded as a new event by Comet,
which will forward it. Thus, we are more open to the risks of infinte event
loops.