Read passphrase from a file, rather than cmd line.
It's not safe to have the user type their passphrase on the command line --
other users on the same system could see it. Instead, they specify a file
name, and we read the passphrase from that. The file should only be readable
by the user running comet, but we don't actually enforce that.