Disable use of gpg-agent
It seems that gpgme (on some platforms?) prefers to call gpg-agent rather than
its passphrase callback, even if the latter is defined. gpg-agent, in turn,
pops up a dialogue box asking for a passphrase which the user has probably
already provided to Comet!

By unsetting GPG_AGENT_INFO in the environment in which gpgme runs, we work
around this.