Options to use UNIX domain sockets for broker or publisher
Add options to use UNIX domain sockets for all of the connection
endpoints available to the broker or publisher.

We operate an instance of comet on a shared system in a computing
cluster to which many users have access. When we were using TCP
sockets, any local user could submit a VOEvent. Adding command
line options to use UNIX domain sockets allows us to control
submission access using filesystem permissions.